ECG SIGNAL PROCESSING
In this project I am doing a simple implementation of a TimeFrequency
analysis on a raw ECG signal, An
ECG signal is obtained from an online
database and it is analyzed using the basic toolbox in Matlab
program and then its features are determined finally its rate is
determined using its features. The ECG signal
is treated with FFT and Wigner Distribution techniques to remove all the
unwanted data in it and the peaks
for each beat are determined, all the results and explanations are done
in the appendix section and also the
code used for this implementation is attached along with the submission.